Add files via upload
This is my initial simulation of a robot football team, including the football field and both players, based on simulink. To run this code you need to first load 'matlab.mat' in matlab and then run the simulink code to achieve the simulation. In addition, the 'Mobile Robotics Simulation Toolbox' needs to be installed to run this code.